+++ hutool-core/src/main/java/cn/hutool/core/thread/AsyncUtil.java
@@ -1,22 +1,17 @@
 package cn.hutool.core.thread;
 
-import cn.hutool.core.exceptions.ExceptionUtil;
-
 import java.lang.reflect.UndeclaredThrowableException;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 
 /**
- * CompletableFuture工具类，叫CompletableFutureUtil太长
+ * {@link CompletableFuture}异步工具类<br>
+ * {@link CompletableFuture} 是 Future 的改进，可以通过传入回调对象，在任务完成后调用之
  *
  * @author <achao1441470436@gmail.com>
- * @since 2021/11/10 0010 20:55
+ * @since 5.7.17
  */
-public class SyncUtil {
-
-	private SyncUtil() {
-		/* Do not new me! */
-	}
+public class AsyncUtil {
 
 	/**
 	 * 等待所有任务执行完毕，包裹了异常
@@ -24,11 +19,25 @@ public class SyncUtil {
 	 * @param tasks 并行任务
 	 * @throws UndeclaredThrowableException 未受检异常
 	 */
-	public static void wait(CompletableFuture<?>... tasks) {
+	public static void waitAll(CompletableFuture<?>... tasks) {
 		try {
 			CompletableFuture.allOf(tasks).get();
 		} catch (InterruptedException | ExecutionException e) {
-			ExceptionUtil.wrapAndThrow(e);
+			throw new ThreadException(e);
+		}
+	}
+
+	/**
+	 * 等待任意一个任务执行完毕，包裹了异常
+	 *
+	 * @param tasks 并行任务
+	 * @throws UndeclaredThrowableException 未受检异常
+	 */
+	public static void waitAny(CompletableFuture<?>... tasks) {
+		try {
+			CompletableFuture.anyOf(tasks).get();
+		} catch (InterruptedException | ExecutionException e) {
+			throw new ThreadException(e);
 		}
 	}
 
@@ -41,13 +50,11 @@ public class SyncUtil {
 	 * @throws RuntimeException 未受检异常
 	 */
 	public static <T> T get(CompletableFuture<T> task) {
-		RuntimeException exception;
 		try {
 			return task.get();
 		} catch (InterruptedException | ExecutionException e) {
-			exception = ExceptionUtil.wrapRuntime(e);
+			throw new ThreadException(e);
 		}
-		throw exception;
 	}
 
 }