+++ hutool-core/src/test/java/cn/hutool/core/thread/AsyncUtilTest.java
@@ -1,6 +1,7 @@
 package cn.hutool.core.thread;
 
 import cn.hutool.core.lang.Assert;
+import org.junit.Ignore;
 import org.junit.Test;
 
 import java.util.concurrent.CompletableFuture;
@@ -12,25 +13,26 @@ import java.util.concurrent.TimeUnit;
  * @author <achao1441470436@gmail.com>
  * @since 2021/11/10 0010 21:15
  */
-public class SyncUtilTest {
+public class AsyncUtilTest {
 
 	@Test
+	@Ignore
 	public void waitAndGetTest() {
 		CompletableFuture<String> hutool = CompletableFuture.supplyAsync(() -> {
-			ThreadUtil.sleep(3, TimeUnit.SECONDS);
+			ThreadUtil.sleep(1, TimeUnit.SECONDS);
 			return "hutool";
 		});
 		CompletableFuture<String> sweater = CompletableFuture.supplyAsync(() -> {
-			ThreadUtil.sleep(4, TimeUnit.SECONDS);
+			ThreadUtil.sleep(2, TimeUnit.SECONDS);
 			return "卫衣";
 		});
 		CompletableFuture<String> warm = CompletableFuture.supplyAsync(() -> {
-			ThreadUtil.sleep(5, TimeUnit.SECONDS);
+			ThreadUtil.sleep(3, TimeUnit.SECONDS);
 			return "真暖和";
 		});
 		// 等待完成
-		SyncUtil.wait(hutool, sweater, warm);
+		AsyncUtil.waitAll(hutool, sweater, warm);
 		// 获取结果
-		Assert.isTrue("hutool卫衣真暖和".equals(SyncUtil.get(hutool) + SyncUtil.get(sweater) + SyncUtil.get(warm)));
+		Assert.isTrue("hutool卫衣真暖和".equals(AsyncUtil.get(hutool) + AsyncUtil.get(sweater) + AsyncUtil.get(warm)));
 	}
 }